+++ krystaldrop/Sources/KDpp/Video/SpriteInstance.h
@@ -129,12 +129,12 @@ public:
 		Displays the image in (x,y).
 		Note: only call a Display function once a frame!
 	*/
-	virtual void Display(int x, int y);
+	virtual void Display(float x, float y);
 
 	/**
 		Displays the image in (x,y) with some alpha blending
 	*/
-	virtual void DisplayAlpha(int x, int y, int alpha);
+	virtual void DisplayAlpha(float x, float y, int alpha);
 
 	/**
 		Displays the image turned around (1,0,0) at (x,y)
@@ -151,7 +151,7 @@ public:
 		This is only supported by the OpenGL renderer.
 		The SDL renderer won't perform zoom nor coloring
 	*/
-	virtual void DisplayColorZoom(int x, int y, int r, int g, int b, int alpha, float resizeX, float resizeY);
+	virtual void DisplayColorZoom(float x, float y, int r, int g, int b, int alpha, float resizeX, float resizeY);
 
 	/**
 		Displays the image with some alpha blending, zooming and coloring and rotation.
@@ -159,23 +159,23 @@ public:
 		This is only supported by the OpenGL renderer.
 		The SDL renderer won't perform zoom nor coloring
 	*/
-	virtual void DisplayColorZoomRotate(int x, int y, int r, int g, int b, int alpha, float resizeX, float resizeY, int rotX, int rotY, float angle);
+	virtual void DisplayColorZoomRotate(float x, float y, int r, int g, int b, int alpha, float resizeX, float resizeY, int rotX, int rotY, float angle);
 
 	/**
 		Displays a part of the image in (x,y) with some alpha blending
 		The part Displayed is delimited by (xPart,yPart,xPart+widthPart,yPart+heightPart)
 	*/
-	virtual void DisplayAlphaPart(int x, int y, int alpha, int xPart, int yPart, int widthPart, int heightPart);
+	virtual void DisplayAlphaPart(float x, float y, int alpha, int xPart, int yPart, int widthPart, int heightPart);
 
 	/**
 		Displays the sprite with full poarameter passing: the coordinates color and alpha of each corner!
 	*/
-	virtual void DisplayFullParameters(int x1, int y1, int r1, int g1, int b1, int alpha1, int x2, int y2, int r2, int g2, int b2, int alpha2, int x3, int y3, int r3, int g3, int b3, int alpha3, int x4, int y4, int r4, int g4, int b4, int alpha4);
+	virtual void DisplayFullParameters(float x1, float y1, int r1, int g1, int b1, int alpha1, float x2, float y2, int r2, int g2, int b2, int alpha2, float x3, float y3, int r3, int g3, int b3, int alpha3, float x4, float y4, int r4, int g4, int b4, int alpha4);
 
 	/**
 		Displays the sprite with full poarameter passing: the coordinates color and alpha of each corner!
 	*/
-	virtual void DisplayFullParameters(int x1, int y1, const KD_Color &c1, int alpha1, int x2, int y2, const KD_Color &c2, int alpha2, int x3, int y3, const KD_Color &c3, int alpha3, int x4, int y4, const KD_Color &c4, int alpha4);
+	virtual void DisplayFullParameters(float x1, float y1, const KD_Color &c1, int alpha1, float x2, float y2, const KD_Color &c2, int alpha2, float x3, float y3, const KD_Color &c3, int alpha3, float x4, float y4, const KD_Color &c4, int alpha4);
 
 	/**
 		Returns the height of the surface.
